Implement `set()`: Add key-value pair into map
`set(key, value)`: Add the given key-value pair into the hashmap.
It uses the hashCode to store that pair into a dedicated bucket.
If bucket is empty, insert to the head of the bucket.
If a head exists in that bucket, add the pair to the end of it.

Feature yet to be added: Triggers hashmap resizing by load factor.