Implement `keys()`: create array of all keys
`keys()` uses `reduce()` to iterate over buckets of the hash map, upon
each iteration, it traverses the linked list and collects every key it
encounters into a list.
Finally, it merges the list to the accumulator: `arr`.
After it finishes reducing, `keys()` returns the completed `arr` that
contains all keys.